style: delete unnecessary accept/reject modal from appliedJob (use from notification page instead)
