--- apps/web/src/modules/job/appliedJob/components/AcceptOfferModal/types.ts
@@ -1,6 +1,0 @@
-export interface AcceptOfferModalProps {
-  isOpen: boolean
-  title: string
-  handleClose(): void
-  handleSubmit(): Promise<void>
-}
