--- apps/web/src/modules/job/appliedJob/components/RejectOfferModal/styled.ts
@@ -1,18 +1,0 @@
-import { Modal, styled } from '@mui/material'
-
-export const RejectOfferModalRootContainer = styled(Modal)`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 12px;
-`
-export const RejectOfferModalContentContainer = styled('div')`
-  width: 100%;
-  max-width: 500px;
-  padding: 16px;
-  background-color: white;
-  border-radius: 8px;
-  display: flex;
-  gap: 12px;
-  flex-direction: column;
-`
