--- apps/web/src/modules/job/appliedJob/components/RejectOfferModal/types.ts
@@ -1,6 +1,0 @@
-export interface RejectOfferModalProps {
-  isOpen: boolean
-  title: string
-  handleClose(): void
-  handleSubmit(): Promise<void>
-}
