+++ apps/web/src/modules/notification/components/AcceptOfferModal/types.ts
@@ -1,4 +1,4 @@
-export interface IOfferModalProps {
+export interface AcceptOfferModalProps {
   isOpen: boolean
   title: string
   handleClose(): void
