+++ apps/web/src/modules/notification/components/RejectOfferModal/styled.ts
@@ -1,12 +1,12 @@
 import { Modal, styled } from '@mui/material'
 
-export const RootModalContainer = styled(Modal)`
+export const RejectOfferModalRootContainer = styled(Modal)`
   display: flex;
   justify-content: center;
   align-items: center;
   padding: 12px;
 `
-export const ModalContentContainer = styled('div')`
+export const RejectOfferModalContentContainer = styled('div')`
   width: 100%;
   max-width: 500px;
   padding: 16px;
