+++ apps/web/src/modules/notification/components/RejectOfferModal/types.ts
@@ -1,4 +1,4 @@
-export interface IOfferModalProps {
+export interface RejectOfferModalProps {
   isOpen: boolean
   title: string
   handleClose(): void
